+++ src/Entity/dynamic_entity.h
@@ -7,7 +7,7 @@ class DynamicEntity : public Entity
 {
 private:
     float speed;
-    Vector2<int> direction;
+    Vector2<float> direction;
 
 protected:
     DynamicEntity(const std::string name) : Entity(name) {}
@@ -16,11 +16,11 @@ public:
     virtual ~DynamicEntity() {}
     virtual std::string getType() = 0;
     void move() override;
-    void setDirection(Vector2<int> direction) override;
+    void setDirection(Vector2<float> direction) override;
 
     inline void setSpeed(float s) { speed = s; }
     inline float getSpeed() const { return speed; }
-    inline Vector2<int> getDirection() const override { return direction; }
+    inline Vector2<float> getDirection() const override { return direction; }
 };
 
 #endif
