+++ src/Entity/entity.h
@@ -25,8 +25,8 @@ public:
     virtual ~Entity();
     virtual std::string getType() = 0;
     virtual void move() = 0;
-    virtual void setDirection(Vector2<int> direction) = 0;
-    virtual Vector2<int> getDirection() const = 0;
+    virtual void setDirection(Vector2<float> direction) = 0;
+    virtual Vector2<float> getDirection() const = 0;
 
     void setPosition(Vector2<int> v);
     void setGraphic(SDL_Rect rect);
