+++ src/Entity/static_entity.h
@@ -12,8 +12,8 @@ public:
     virtual ~StaticEntity() {}
     virtual std::string getType() = 0;
     inline void move() override {}
-    inline void setDirection(Vector2<int> direction) override {}
-    inline Vector2<int> getDirection() const override { return Vector2<int>(0, 0); }
+    inline void setDirection(Vector2<float> direction) override {}
+    inline Vector2<float> getDirection() const override { return Vector2<float>(0, 0); }
 };
 
 #endif
